+++ SuyoStore/Assets/1.Scripts/Manager/SoundController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 012c001cf471fd74f9f46c61265bd514
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
