+++ SuyoStore/Assets/1.Scripts/Manager/SoundManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 012c001cf471fd74f9f46c61265bd514
+guid: 451c15c6e32faf04d853145654b81e6d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
