--- SuyoStore/Assets/Prefabs/Debuff.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ba59a302e83586a42bd370709805986c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
