+++ SuyoStore/Assets/Prefabs/GameManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a4706cae07a6dd74683a8d751c60427f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
