--- SuyoStore/Assets/Prefabs/Item List.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d73e5c729256228449a1aa14ac4043ea
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
