+++ SuyoStore/Assets/Prefabs/Sound Manager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 69a492bae5ed4c441be6d837cd00f663
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
