+++ SuyoStore/Assets/Prefabs/UI Manager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fe433134125f4654f9896dfa0d56b199
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
