+++ SuyoStore/Assets/Prefabs/UI/Debuff.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ba59a302e83586a42bd370709805986c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
