+++ SuyoStore/Assets/Prefabs/UI/Item List.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d73e5c729256228449a1aa14ac4043ea
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
