+++ SuyoStore/Assets/Sounds/Test.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2b48b7e1fbe60f0408e43aecd0a93c65
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
